requirements.txt
mysqlclient==2.0.3
Django==3.2.7
django-filter==21.1
djangorestframework==3.12.4